File Browser Adjusted
PLEASE check this out... not sure if this is done right...
Changed `md-sidenav` to a div with the same classes, and made it's
minimum width 275px in the scss. it looks good right now... when the
viewport width shrinks, it doesn't at all, but it works for now...
MAY/be it SHOULD be a sidenav still? not sure.

I also shifted `md-content` tags around a little, this makes the div
scroll nicely no matter how many txt files are on display.

I also added the hover effect

I also removed file icon

I also made md-list-items '2 line' list items.

I also moved the uploadButton outside of the `md-content` so it won't
scroll.