Trimmed Notes, BROKE Something.
Hope it is an easy enough fix...  Broke it with my previous commit.

Now when clicking a file, it doesn't load the meta-data (or the file...
But i don't think it loaded the file anyways...).

Pretty sure it has to do with `ng-click="null"` in the `md-list-item`,
but i can't find another way to make the hover effect work otherwise.

Go ahead an revert any changes... It is all good.